Changes:
BREAKING CHANGE: The files from LinFu.IoC.Extensions and LinFu.IoC.Plugins have been moved into the LinFu.IoC.Configuration namespace. This change was necessary because the loader class needs to be able to load the default dependencies into a container and that wasn't possible because of the diamond dependency issues that LinFu.IoC.Extensions and LinFu.IoC.Plugins were causing, and thus, they had to be removed.